scranton/agents/holly_flax/requirements.txt
google-cloud-aiplatform[adk,agent-engines]>=1.97.0
python-dotenv>=1.1.0
litellm>=1.72.7